requirements_clean.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.26.0
anyio==4.3.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
defusedxml==0.7.1
distro==1.9.0
docutils==0.21.2
executing==2.0.1
Faker==25.2.0
fastapi==0.109.2
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
fvalues==0.0.4
git-filter-repo==2.38.0
google-ai-generativelanguage==0.6.4
google-api-core==2.17.1
google-api-python-client==2.129.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
google-search-results==2.4.2
googleapis-common-protos==1.63.0
gorilla==0.4.0
GPUtil==1.4.0
gradio==4.31.4
gradio_client==0.16.4
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.6
importlib-resources==6.1.1
inquirerpy==0.3.4
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.0
langchain-anthropic==0.1.13
langchain-community==0.0.38
langchain-core==0.2.0
langchain-google-genai==1.0.4
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchainhub==0.1.15
langsmith==0.1.57
Levenshtein==0.25.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
mdurl==0.1.2
merge-args==0.1.5
more-itertools==10.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numerizer==0.2.3
numexpr==2.9.0
numpy==1.26.4
onnxruntime-genai-directml==0.2.0rc6
openai==1.30.1
orjson==3.9.14
ought-ice==0.5.0
packaging==23.2
pandas==2.2.0
pathlib==1.0.1
pfzy==0.3.4
pillow==10.2.0
pockets==0.9.1
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
python-ulid==2.5.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
questionary==2.0.1
rapidfuzz==3.9.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rouge-metric==1.0.1
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.2.2
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sphinxcontrib-napoleon==0.7
SQLAlchemy==2.0.28
starlette==0.36.3
structlog==24.1.0
tenacity==8.2.3
text-generation==0.7.0
thefuzz==0.22.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.2
transformers==4.39.2
typer==0.12.3
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.27.1
wcwidth==0.2.13
websockets==11.0.3
wheel==0.41.2
yarl==1.9.4